pyproject.toml
[tool.poetry]
name = "gmail-spam"
version = "0.1.0"
description = ""
authors = ["Maja <majam1616@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.19.2"
boto3 = "^1.21.23"
imaplib2 = "^3.6"

[tool.poetry.dev-dependencies]
moto = {extras = ["s3"], version = "^3.1.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
